refactor: extract component `ImageCarousel`
