+++ src/data/home.ts
@@ -1,0 +1,16 @@
+import type { Image } from '@/components/ImageCarousel';
+
+export const CAROUSEL_IMAGES: Image[] = [
+    {
+        src: '/images/home/duck-ctf.jpg',
+        alt: 'DuckCTF',
+    },
+    {
+        src: '/images/home/pizza.jpg',
+        alt: 'Pizza',
+    },
+    {
+        src: '/images/home/cyber-panel.jpg',
+        alt: 'Cyber Panel',
+    },
+];
